requirements.txt
flask
requests
python-telegram-bot==20.3
fuzzywuzzy
python-Levenshtein
gTTS
SpeechRecognition
pydub
whois
